Moved NotNull to FieldSerializer.NotNull.
It only applies to FieldSerializer and has always been in the wrong place.
